+++ datacollection/src/main/java/org/opendaylight/tsdr/datacollection/listeners/NodeConnectorStatisticsChangeListener.java
@@ -98,7 +98,7 @@ public class NodeConnectorStatisticsChangeListener extends
                         .getTSDRMetricRecordBuilderContainer(id);
                 if (bc != null) {
                     TSDRMetricRecordBuilder builder[] = bc.getBuilders();
-                    BigInteger timeStamp = getTimeStamp();
+                    long timeStamp = getTimeStamp();
                     builder[0].setMetricValue(new Counter64(new BigInteger(""
                             + fs.getTransmitDrops())));
                     builder[0].setTimeStamp(timeStamp);
