+++ datacollection/src/main/java/org/opendaylight/tsdr/datacollection/listeners/NodeGroupStatisticsChangeListener.java
@@ -72,7 +72,7 @@ public class NodeGroupStatisticsChangeListener extends
                         .getTSDRMetricRecordBuilderContainer(id);
                 if (bc != null) {
                     TSDRMetricRecordBuilder builder[] = bc.getBuilders();
-                    BigInteger timeStamp = getTimeStamp();
+                    long timeStamp = getTimeStamp();
                     builder[0].setMetricValue(new Counter64(new BigInteger(""
                             + gs.getRefCount())));
                     builder[0].setTimeStamp(timeStamp);
